I2S regfile: Set wb_data_o to 0 when not reading
This is not strictly required in wishbone, the master should just ignore wb_data_o as long as we're not in a stb or read. However, we want to reuse the NEORV32 bus switch, so we need to support the internal protocol which is a little more strict. Here we should always set the output to 0 when we're not being addressed.