Externalize memory for sfifo. Use single port memory
Give priority to read. The write port has to stall if there's a parallel read. Introduce a new o_ready signal to show this and make use of it in the reg file.

Note: This does currently not work with unpatched xbus, see https://github.com/stnolting/neorv32/discussions/1290

Patch will be included in PSOC NEORV fork